TLS: Remove TLS types
TLS Types are no longer required in utils and instead are relocated to
juju/http
